mj-bot/pyproject.toml
[project]
name = "mj-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "discord.py-self>=2.0.0",
    "python-dotenv>=1.2.1",
    "websockets>=16.0",
]
